Fix textinput focus issues after updating to api11
closes #729 

After API 11 update the focus logic in TextInput had to be updated, because the new TextInput will not focus automatically, even with `defaultFocus: true` as this controls the focus only when the page is first rendered - and the whole RNApp is generally rendered on a singular ArkUI page. The `autoFocus` property is now realised by manually focusing the TextInput in `onAppear` callback when needed.

This also fixes the bug where the keyboard blinks when using the descriptor tree is rearranged, causing the TextInput to briefly rerender before dissapearing.

See merge request rnoh/react-native-harmony!687