Fix image rendering inside the bordered area, not below it
Previously when a border was set on the image, the image would render inside the bordered area, resizing it to fit which was inconsistent with other platforms. Now it is rendered in the whole available parent space and the borders cover the outermost parts of the image.

This also fixes a bug, where if borderColor would not be set then the border would be transparent instead of black as expected.

closes #335

See merge request rnoh/react-native-harmony!699