Change tabIndex test to make it passable on Harmony devices
`tabIndex` is resolved in JS to `focusable` so it was already working. Refactored the test to make it work also on Harmony devices, as the previous one relied on the visual response from focusing, and this is not available in ArkUI. 

`@ts-ignore` flag is needed as `tabIndex` is not exported in RN for some reason.

closes #258

See merge request rnoh/react-native-harmony!719