Fix height/width not being set correctly when 0
## Summary

When `height` or `width` of the component were set to `0` the resulting size would be determined by the size of its children because `maybeAssignAttribute` would compare this values to the provided default - `0` - which would end in not setting the attribute. However the default value for a container size in ArkUI is not `0` but rather `undefined` - if not provided it will enlarge to fit its children.

## Changes

* changed default values for `width` and `height` from `0` to `undefined` in `RNViewBaseAttributeModifier` 
* added a test case for that problem

## Test Plan

1. Open tester app.
2. Navigate to `ViewTest` \> `"not show view with height 0 and overflow hidden"` test case
3. Observe that the `View` is invisible

## Checklist

- [x] Project compiles and the tester application works.

## Issue Links

closes #891

See merge request rnoh/react-native-harmony!899