Implement missing ScrollView features
## Summary

This MR adds implementations of some of the missing properties of `ScrollView`. It is loosely based on !925. Due to bugs in ArkUI some of the features have certain limitations

* `snapToOffsets` with only one offset given will behave as `snapToInterval` because of how ArkUI interprets arguments passed to `NODE_SCROLL_SNAP` attribute
* `snapToStart` and `snapToEnd` is always `true` when `snapToOffsets` is used 
* both `scrollTo` and `scrollToEnd` commands with `animated: true` set will not execute because of ArkUI bugs

Additionally, `flashScrollIndicators` is implemented as a patch on the JS side, since there is no mechanism yet for properly setting a timeout for hiding the scroll bar.

## Changes

* added `scrollTo` and `scrollToEnd` commands
* added `scrollBar` specific properties
* added `snapToOffsets`, `snapToInterval`
* added `pagingEnabled`
* added `scrollEventThrottle`
* added `flashScrollIndicators` command

## Test Plan

1. Open tester app.
2. Observe `ScrollView` related tests working as expected

## Checklist

- [x] Project compiles and the tester application works.

## Issue Links

See merge request rnoh/react-native-harmony!930