fix: align scroll view content to top-left [C-API]
## Summary
<!--
Briefly describe the purpose of the MR, for example:

This PR adds support for rounded corners for components in C-API architecture. 
-->
ArkUI aligns `ScrollView` content to the center by default, while RN expects it to be aligned to the start of the `ScrollView`. This PR sets the alignment of the `Scroll` node to take that into account.

## Changes
<!-- List main changes in bullet points that can be included in the changelog. E.g.,
- Added rounded corner support in C-API architecture.
-->

## Test Plan
<!-- Steps on how the changes were tested. E.g.,
1. Open tester app.
2. Run "ViewTest::render square with rounded corners with different radii (left/right)". -->

1. Comment out test suites so that the content of tester is shorter than the screen.
1. Open tester app.
1. Verify the content is aligned to the top of the screen, not the center.

## Checklist
- [x] Project compiles and the tester application works.

## Issue Links
Closes #907

See merge request rnoh/react-native-harmony!938