Provide better nested scrolling experience
## Summary

ArkUI's `NestedScrollMode.SELF_FIRST`, which was previously used to handle propagation of nested scrolling to parent, does not stop scrolling when the child reaches its end. This made it easy to scroll all the way to the top of the outermost scroll with just one, strong fling. By handling it manually we make sure that when a child is not at its end when the scrolling started the parent will not be scrolled, no matter how fast/strong the fling is.

## Changes

* nested scrolling is now handled manually
  * when a child Scroll is scrolled from its end towards the outside, the parent Scroll starts scrolling
  * when a child Scroll is scrolled from its end towards the inside or it is scrolled from its inside, the parent Scroll will not start scrolling when the child's end is reached
* bouncing is now turned off in nested Scrolls whose content size is smaller than container size (just like on iOS)

## Test Plan

1. Open tester app.
2. Go to `NestedScrollingExample` 
3. Test nested scrolling manually by scrolling the child 
   * from its end towards the outside
   * from its end towards the inside
   * from its inside up or down

## Checklist

- [x] Project compiles and the tester application works.

## Issue Links

See merge request rnoh/react-native-harmony!1018