Additional tests for multiple folders images support
## Summary
<!--
Briefly describe the purpose of the MR, for example:

This PR adds support for rounded corners for components in C-API architecture. 
-->

This PR adds additional test for Images. After investigating an issue reported on Discord channel I created this test for the future testing purpose. The goal of the test is to check if loading images assets from different subfolder (and nested subfolders) works correctly.

## Changes
<!-- List main changes in bullet points that can be included in the changelog. E.g.,
- Added rounded corner support in C-API architecture.
-->

- Added additional test for loading images from different subfolders.

## Test Plan
<!-- Steps on how the changes were tested. E.g.,
1. Open tester app.
2. Run "ViewTest::render square with rounded corners with different radii (left/right)". -->

1. Open tester app.
2. Navigate to the Image Tab and look for the `Load images from different folders within assets directory (e.g. "assets/large-images", "assets/small-images")`


## Checklist
- [x] Project compiles and the tester application works.

## Issue Links
<!-- Link related or resolved issues. E.g.,
Related: #123 
Closes: #456
-->

See merge request rnoh/react-native-harmony!1023