fix: textinput clearButtonMode not support while-editing
## Summary
<!--
Briefly describe the purpose of the MR, for example:

This PR adds support for rounded corners for components in C-API architecture. 
-->

This PR adds support for textInput component clearButtonMode property setting for 'while-editing'

## Changes
<!-- List main changes in bullet points that can be included in the changelog. E.g.,
- Added rounded corner support in C-API architecture.
-->

add check to see if the textInput set clearButtonMode as 'while-editing', if yes, display the clear button while there is input in textInput. 

## Test Plan
<!-- Steps on how the changes were tested. E.g.,
1. Open tester app.
2. Run "ViewTest::render square with rounded corners with different radii (left/right)". -->

1. Open tester app.
2. check on clearButtonMode with setting 'while-editing'


## Checklist
- [x] Project compiles and the tester application works.

## Issue Links
<!-- Link related or resolved issues. E.g.,
Related: #123 
Closes: #456
-->

See merge request rnoh/react-native-harmony!1037